рефакторинг скрипта формы редактирования
